+++ ouroboros-network-api/src/Ouroboros/Network/PeerSelection/LocalRootPeers.hs
@@ -1,0 +1,5 @@
+module Ouroboros.Network.PeerSelection.LocalRootPeers (OutboundConnectionsState (..)) where
+
+data OutboundConnectionsState = ConnectedToExternalOutboundPeers
+                              | ConnectedToOnlyLocalOutboundPeers
+  deriving (Eq, Show)